+++ .eslintrc.json
@@ -1,6 +1,6 @@
 {
   "extends": [
-    "next",
+    "next/core-web-vitals",
     "prettier"
   ]
 }
