--- app/news/[slug]/(.)image/page.js
@@ -1,23 +1,0 @@
-import { notFound } from "next/navigation";
-
-import { DUMMY_NEWS } from "@/dummy-news";
-
-export default function interceptedImagePage({ params }) {
-const newsItemSlug = params.slug;
-const newsItem = DUMMY_NEWS.find(
-(newsItem) => newsItem.slug === newsItemSlug,
-);
-
-if (!newsItem) {
-notFound();
-}
-
-return (
-<>
-<h2>Intercepted</h2>
-<div className="fullscreen-image">
-<img src={`/images/news/${newsItem.image}`} alt={newsItem.title} />
-</div>
-</>
-);
-}
