+++ app/news/[slug]/@modal/(.)image/page.js
@@ -1,0 +1,25 @@
+import { notFound } from "next/navigation";
+
+import { DUMMY_NEWS } from "@/dummy-news";
+
+export default function interceptedImagePage({ params }) {
+  const newsItemSlug = params.slug;
+  const newsItem = DUMMY_NEWS.find(
+    (newsItem) => newsItem.slug === newsItemSlug,
+  );
+
+  if (!newsItem) {
+    notFound();
+  }
+
+  return (
+    <>
+      <div className="modal-backdrop" />
+      <dialog className="modal" open>
+        <div className="fullscreen-image">
+          <img src={`/images/news/${newsItem.image}`} alt={newsItem.title} />
+        </div>
+      </dialog>
+    </>
+  );
+}
