+++ app/news/[slug]/@modal/page.js
@@ -1,0 +1,3 @@
+export default function ModalDefaultPage() {
+  return null;
+}
