+++ app/news/[slug]/layout.js
@@ -1,0 +1,8 @@
+export default function NewsDetailLayout({ children, modal }) {
+  return (
+    <>
+      {modal}
+      {children}
+    </>
+  )
+}
