+++ package.json
@@ -14,14 +14,8 @@
     "react-dom": "^18"
   },
   "devDependencies": {
-    "@next/eslint-plugin-next": "^14.2.3",
-    "@typescript-eslint/eslint-plugin": "^7.8.0",
-    "@typescript-eslint/parser": "^7.8.0",
     "eslint": "^8.57.0",
     "eslint-config-next": "14.2.3",
-    "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-prettier": "^5.1.3",
-    "prettier": "^3.2.5",
-    "prettier-plugin-tailwindcss": "^0.5.14"
+    "eslint-config-prettier": "^9.1.0"
   }
 }
